[department-of-veterans-affairs/orchid#190] combined ask/asks and ask/inquiries into single inquiries controller
