+++ app/controllers/v0/apidocs_controller.rb
@@ -122,8 +122,7 @@ module V0
       Swagger::Requests::Address,
       Swagger::Requests::Appeals::Appeals,
       Swagger::Requests::Appointments,
-      Swagger::Requests::ContactUs::Asks,
-      Swagger::Requests::ContactUs::InquiriesList,
+      Swagger::Requests::ContactUs::Inquiries,
       Swagger::Requests::BackendStatuses,
       Swagger::Requests::BB::HealthRecords,
       Swagger::Requests::BurialClaims,
@@ -177,7 +176,7 @@ module V0
       Swagger::Schemas::Address,
       Swagger::Schemas::Appeals::Requests,
       Swagger::Schemas::Appeals::HigherLevelReview,
-      Swagger::Schemas::ContactUs::Asks,
+      Swagger::Schemas::ContactUs::SuccessfulInquiryCreation,
       Swagger::Schemas::ContactUs::InquiriesList,
       Swagger::Schemas::AsyncTransaction::Vet360,
       Swagger::Schemas::BB::HealthRecords,
