--- app/controllers/v0/contact_us/asks_controller.rb
@@ -1,38 +1,0 @@
-# frozen_string_literal: true
-
-module V0
-  module ContactUs
-    class AsksController < ApplicationController
-      skip_before_action :authenticate, only: :create
-
-      def create
-        return not_implemented unless Flipper.enabled?(:get_help_ask_form)
-
-        claim = SavedClaim::Ask.new(form: form_submission)
-
-        validate!(claim)
-
-        render json: {
-          'confirmationNumber': '0000-0000-0000',
-          'dateSubmitted': DateTime.now.utc.strftime('%m-%d-%Y')
-        }, status: :created
-      end
-
-      private
-
-      def not_implemented
-        render nothing: true, status: :not_implemented, as: :json
-      end
-
-      def form_submission
-        params.require(:inquiry).require(:form)
-      rescue
-        raise
-      end
-
-      def validate!(claim)
-        raise Common::Exceptions::ValidationErrors, claim unless claim.valid?
-      end
-    end
-  end
-end
