+++ app/controllers/v0/contact_us/inquiries_controller.rb
@@ -3,10 +3,41 @@
 module V0
   module ContactUs
     class InquiriesController < ApplicationController
+      skip_before_action :authenticate, only: :create
+
       def index
         return not_implemented unless Flipper.enabled?(:get_help_messages)
 
         render json: STUB_RESPONSE, status: :ok
+      end
+
+      def create
+        return not_implemented unless Flipper.enabled?(:get_help_ask_form)
+
+        claim = SavedClaim::Ask.new(form: form_submission)
+
+        validate!(claim)
+
+        render json: {
+            'confirmationNumber': '0000-0000-0000',
+            'dateSubmitted': DateTime.now.utc.strftime('%m-%d-%Y')
+        }, status: :created
+      end
+
+      private
+
+      def form_submission
+        params.require(:inquiry).require(:form)
+      rescue
+        raise
+      end
+
+      def validate!(claim)
+        raise Common::Exceptions::ValidationErrors, claim unless claim.valid?
+      end
+
+      def not_implemented
+        render nothing: true, status: :not_implemented, as: :json
       end
 
       STUB_RESPONSE = {
@@ -49,12 +80,6 @@ module V0
           }
         ]
       }.freeze
-
-      private
-
-      def not_implemented
-        render nothing: true, status: :not_implemented, as: :json
-      end
     end
   end
 end
